Moved out of the __init__.py the functions: rate_this_addon()
